Removed Nav/Footer from core layout + renamed demo layout SSG/SSR utils
