+++ src/pages/[locale]/demo/built-in-features/analytics.tsx
@@ -10,7 +10,7 @@ import DemoPage from '@/layouts/demo/components/DemoPage';
 import {
   getDemoStaticPaths,
   getDemoStaticProps,
-} from '@/layouts/demo/demoSSG';
+} from '@/layouts/demo/demoLayoutSSG';
 import { LogEvent } from '@/modules/core/amplitude/types/Amplitude';
 import { createLogger } from '@/modules/core/logging/logger';
 import useUserConsent from '@/modules/core/userConsent/hooks/useUserConsent';
