+++ src/pages/[locale]/demo/built-in-features/docs-site.tsx
@@ -8,7 +8,7 @@ import DemoPage from '@/layouts/demo/components/DemoPage';
 import {
   getDemoStaticPaths,
   getDemoStaticProps,
-} from '@/layouts/demo/demoSSG';
+} from '@/layouts/demo/demoLayoutSSG';
 import { createLogger } from '@/modules/core/logging/logger';
 import {
   GetStaticPaths,
