+++ src/pages/[locale]/demo/built-in-utilities/errors-handling.tsx
@@ -9,7 +9,7 @@ import DemoPage from '@/layouts/demo/components/DemoPage';
 import {
   getDemoStaticPaths,
   getDemoStaticProps,
-} from '@/layouts/demo/demoSSG';
+} from '@/layouts/demo/demoLayoutSSG';
 import I18nLink from '@/modules/core/i18n/components/I18nLink';
 import { createLogger } from '@/modules/core/logging/logger';
 import {
