+++ src/pages/[locale]/demo/built-in-utilities/top-level-500-error.tsx
@@ -6,7 +6,7 @@ import DemoLayout from '@/layouts/demo/components/DemoLayout';
 import {
   getDemoStaticPaths,
   getDemoStaticProps,
-} from '@/layouts/demo/demoSSG';
+} from '@/layouts/demo/demoLayoutSSG';
 import { createLogger } from '@/modules/core/logging/logger';
 import { isBrowser } from '@unly/utils';
 import {
