+++ src/pages/[locale]/demo/native-features/example-with-ssg-and-revalidate.tsx
@@ -11,7 +11,7 @@ import NativeFeaturesSidebar from '@/layouts/demo/components/NativeFeaturesSideb
 import {
   getDemoStaticPaths,
   getDemoStaticProps,
-} from '@/layouts/demo/demoSSG';
+} from '@/layouts/demo/demoLayoutSSG';
 import { initializeApollo } from '@/modules/core/apollo/apolloClient';
 import { Product } from '@/modules/core/data/types/Product';
 import timeDifference from '@/modules/core/date/timeDifference';
